Move migration test to endpoint test file
